Remove redundant pages
